+++ main.js
@@ -2,12 +2,23 @@
 let knowledge = 0;
 let depression = 0;
 let baseIncrease = 1;
-let depressionStatus = null;
+let depressionStatus = "Ok";
 
+/*
+Ideas:
+
+- Add when you get 100 knowledge you get awareness of your death time, showing an age timer
+- When you get to age 50 you get a disease and die
+- You CAN prevent this by getting higher VIT, which is unlocked by paying 100 knowledge
+
+
+
+*/
 
 // Knowledge accumilination system
 document.getElementById("createKnowledge").addEventListener("click", () => {
     knowledge += baseIncrease;
+
     updateKnowledge();
 });
 
@@ -17,6 +28,7 @@ function updateKnowledge() {
     document.getElementById("displayKnowledge").innerText = "Knowledge: " + knowledge;
 };
 
+
 // Chatbox creation to send a message
 
 function say(message) {
@@ -24,9 +36,9 @@ function say(message) {
 };
 
 
-// Depression Mechanics
+// Depression Mechanics + time
 setInterval(() => {
-    depression += (knowledge/100-10)
+    depression += (knowledge/10-10)
 
     if (depression>99) {
         depressionStatus = "Death awaits, calmly."
@@ -47,24 +59,23 @@ setInterval(() => {
     
 
     document.getElementById("displayDepression").innerText = "Mood: " + depressionStatus;
-}, 100);
+
+    updateTime();
+}, 1000);
     
 // Time System
 let day = 0;
 let year = 14;
 
-setInterval(() => {
-    day += 1;
-    updateTime();
-    document.getElementById("displayDay").innerText = "Day: " + day;
-    document.getElementById("displayYear").innerText = "Year: " + year;
-}, 1000);
-
 function updateTime() {
+    day += 1;
     if (day > 365) {
         day = 0;
         year += 1;
     };
+
+    document.getElementById("displayDay").innerText = "Day: " + day;
+    document.getElementById("displayYear").innerText = "Age: " + year;
 };